allow specification of scan targets through a local file Allow for the specification of scan targets through special hostsfiles. This change adds a file discovery mechanism that will load static targets from a yaml based hosts file on disk.
A hosts file is broken up into one or more target groups, with each group containing a list of hosts that should be parsed into scan targets.

For example, this is taken from the helm chart and will create targets in two separate groups
```yaml
 hosts.yaml: |
    groups:
    - hosts:
      - host: https://vanity.somecompany.com
      - host: https://www.somecompany.com
      - host: 10.2.3.4:8443
      source: important company fqdns
    - hosts:
      - host: https://google.com
      - host: https://github.com
      - host: https://golang.org
      source: other urls
```
Up unitl now scans were against an ip and port but here a URLAddress is introduced so that we can deal with fqdns. If an https scheme is specifed then 443 will be scanned, otherwise a port is required. The config file format is pretty simplistic to begin with, but the intention is to add other configuration like timeouts and additional labels for reporting.